Problem: account tx don't update account states
Solution:
- added a basic implementation in deliver_tx and commit
- moved ABCI-level tests from mod.rs to a separate file
- returning updated account state in validation
- encoding/decoding TX with outputs for validation
