+++ chain-abci/src/app/commit.rs
@@ -1,14 +1,30 @@
 use super::ChainNodeApp;
-use crate::storage::tx::update_utxos_commit;
+use crate::app::spend_utxos;
+use crate::storage::tx::TxWithOutputs;
 use crate::storage::*;
 use abci::*;
+use bit_vec::BitVec;
 use chain_core::common::MerkleTree;
 use chain_core::compute_app_hash;
+use chain_core::tx::data::Tx;
 use chain_core::tx::data::TxId;
 use chain_core::tx::TransactionId;
 use chain_core::tx::TxAux;
 use integer_encoding::VarInt;
+use kvdb::{DBTransaction, KeyValueDB};
 use parity_codec::Encode;
+use std::sync::Arc;
+
+/// Given a db and a DB transaction, it will go through TX inputs and mark them as spent
+/// in the TX_META storage and it will create a new entry for TX in TX_META with all outputs marked as unspent.
+pub fn update_utxos_commit(tx: &Tx, db: Arc<dyn KeyValueDB>, dbtx: &mut DBTransaction) {
+    spend_utxos(&tx.inputs, db, dbtx);
+    dbtx.put(
+        COL_TX_META,
+        &tx.id(),
+        &BitVec::from_elem(tx.outputs.len(), false).to_bytes(),
+    );
+}
 
 impl ChainNodeApp {
     /// Commits delivered TX: flushes updates to the underlying storage
@@ -25,14 +41,43 @@ impl ChainNodeApp {
                 .collect();
             let tree = MerkleTree::new(ids);
             for txaux in self.delivered_txs.iter() {
-                match txaux {
+                let txid: TxId = txaux.tx_id();
+                match &txaux {
                     TxAux::TransferTx(tx, witness) => {
-                        let txid: TxId = tx.id();
-                        inittx.put(COL_BODIES, &txid[..], &tx.encode());
+                        inittx.put(
+                            COL_BODIES,
+                            &txid[..],
+                            &TxWithOutputs::Transfer(tx.clone()).encode(),
+                        );
                         inittx.put(COL_WITNESS, &txid[..], &witness.encode());
                         update_utxos_commit(&tx, self.storage.db.clone(), &mut inittx);
                     }
-                    _ => unimplemented!("MUST_TODO -- account-related TX commits"),
+                    TxAux::DepositStakeTx(tx, witness) => {
+                        inittx.put(COL_BODIES, &txid[..], &tx.encode());
+                        inittx.put(COL_WITNESS, &txid[..], &witness.encode());
+                        // this is not necessary (as they are spent in deliver_tx) and more of a sanity check (as update_utxos_commit does it)
+                        spend_utxos(&tx.inputs, self.storage.db.clone(), &mut inittx);
+                        // account should be already updated in deliver_tx
+                    }
+                    TxAux::UnbondStakeTx(tx, witness) => {
+                        inittx.put(COL_BODIES, &txid[..], &tx.encode());
+                        inittx.put(COL_WITNESS, &txid[..], &witness.encode());
+                        // account should be already updated in deliver_tx
+                    }
+                    TxAux::WithdrawUnbondedStakeTx(tx, witness) => {
+                        inittx.put(
+                            COL_BODIES,
+                            &txid[..],
+                            &TxWithOutputs::StakeWithdraw(tx.clone()).encode(),
+                        );
+                        inittx.put(COL_WITNESS, &txid[..], &witness.encode());
+                        // account should be already updated in deliver_tx
+                        inittx.put(
+                            COL_TX_META,
+                            &txid[..],
+                            &BitVec::from_elem(tx.outputs.len(), false).to_bytes(),
+                        );
+                    }
                 }
             }
             new_state.rewards_pool.last_block_height = new_state.last_block_height;
@@ -43,6 +88,7 @@ impl ChainNodeApp {
             );
             inittx.put(COL_MERKLE_PROOFS, &app_hash[..], &tree.encode());
             new_state.last_apphash = app_hash;
+            new_state.last_account_root_hash = self.uncommitted_account_root_hash;
         }
 
         inittx.put(
